Update: Use sanitizeFolder function
